fixed idefinite value problem (enble needed to be added to brams for 64 to 256 convert). Had to add munge block to output of sp256_data as order is changed in the unpack in the feng. This solved spead issues.
NOTE: an upack error still exists for spead in the receiver. The pkt_word_ctr rst is held
high for 3 cycles to allow count to value of 160 for pkt length count. The problem is that
it is driven by a register that is armed by an eof - but no eof isprsent to arm it for the initial
packet (and header info). This causes a pkt_ln_err right in the beginning for the first packet.
A fix is needed.